Added the ability for the player to "talk" to the buildings
In the player.py file (lines 63-80 as of now) when the player presses the enter/return key, the game will check if they are in the area in front of a building. 

If they are, as of now only a message will be printed, but the plan is to pull up a menu for that building.